requirements.txt
django
djangorestframework
djangorestframework-simplejwt
psycopg2
redis
requests
uwsgi
